new resources package. moved some files around and changed imports to match.
